Create IterativeDeepening.chpl
Let's generate a Document Management Software search tool written in Chapel programming language that implements search via Iterative deepening depth-first search algorithm across storage arrays that the user selects to run their query. Suggest a file name upon completion. 

Sure, we can certainly provide an example of how this might look in Chapel, a language known for its parallel computing capabilities. The Iterative Deepening Depth-First Search (IDDFS) algorithm can be implemented for this purpose.

However, the Chapel language, at the time of my knowledge cutoff in September 2021, did not have native support for file system operations like reading directory contents or file metadata. These capabilities would be needed to traverse directory structures and perform a depth-first search on a file system. Chapel can use C Interoperability to use C libraries that allow such operations, but this might not be ideal.

Assuming the required capabilities are present, an outline of how the IDDFS algorithm could be implemented in Chapel would look like this:

This is only an illustrative implementation, and it doesn't involve the specifics of a document management software such as searching files in a directory or storage arrays. In real-life scenario, you'd need to define what constitutes a "node", how to get the children of a node (e.g., files within a directory), and how to determine if a node is the target of your search.