Create IDDFileSearch_ExceptionHandled.cpp
refactor that code to handle possible exceptions that may occur when accessing files or directories (for example, due to lack of permission) to make it more robust.

This code catches std::filesystem::filesystem_error exceptions which can be thrown by the std::filesystem::directory_iterator constructor if it fails to open the directory (e.g., due to insufficient permissions). It prints an error message to the console and continues with the next directory/file.

You will need a C++17 (or later) compatible compiler to compile this script. The compilation command for GCC would be g++ -std=c++17 IDDFileSearch_ExceptionHandled.cpp -o IDDFileSearch_ExceptionHandled.