Create diskSelectSearch.sh
here's a simple bash script that prompts the user to select a disk from a list of mounted filesystems, and then runs the IDDFileSearch_Extension.cpp program on the selected disk.

This script first checks if the IDDFileSearch_Extension executable exists, and if not, it compiles the IDDFileSearch_Extension.cpp file. Then, it uses the lsblk command to get a list of filesystems, and uses grep and awk to filter and format the list. The select command is used to present the list to the user and get their selection. Finally, it prompts the user to enter the other parameters and runs the C++ program with the provided parameters.

This script assumes that the IDDFileSearch_Extension.cpp file is in the same directory and can be compiled with the command g++ -std=c++17 IDDFileSearch_Extension.cpp -o IDDFileSearch_Extension. You may need to modify this command if your build environment is different.

This is a very simple script and it might not work perfectly in all situations. For example, it does not handle spaces in disk names or filesystem paths, and it assumes that all filesystems listed by lsblk are searchable. For a robust solution, you would likely need to add additional error checking and handling code.

You can save this script to a file named diskSelectSearch.sh. Don't forget to give it execution permissions with chmod +x diskSelectSearch.sh command. You can run it with ./diskSelectSearch.sh.

Please note that in the current context, disk and partition are used interchangeably. The lsblk command lists all block devices, and we're filtering for partitions using grep part. You may adjust this as per your requirements.