Create IDDFileSearch_Extension_Concurrent.cpp
Let's refactor the IDDFileSearch_Extension.cpp code we generated by implementing a Divide-and-conquer search algorithm. Suggest a file name upon completion. 

A divide-and-conquer algorithm works by recursively breaking down a problem into two or more sub-problems of the same or related type, until these become simple enough to be solved directly.

Applying divide-and-conquer strategy to your existing code, the main point of divergence would be in the recursive function DLS(). This function would be modified to concurrently search through directories rather than sequentially. To achieve this, we would need to leverage multi-threading, which can be done using C++11's <thread> standard library.

Here's the updated code:

Please note that multi-threading can provide a significant performance improvement if there are a large number of subdirectories to search through, but it also introduces additional complexity and can actually hurt performance if the number of subdirectories is small due to the overhead of creating and managing threads.

You can save this code in a file named IDDFileSearch_Extension_Concurrent.cpp.

Compile and run this script with commands similar to:

g++ -std=c++17 IDDFileSearch_Extension_Concurrent.cpp -o IDDFileSearch_Extension_Concurrent -pthread
./IDDFileSearch_Extension_Concurrent .txt / 5

Here, .txt is the target file extension, / is the root path, and 5 is the maximum search depth. -pthread flag is required for using multi-threading features.