Create BFSFileSearch.cpp
Refactoring the existing code to use a Breadth-First Search (BFS) algorithm involves changing the search strategy. In BFS, we explore all the files and directories in the current directory before moving on to the subdirectories. Here's how you could modify the existing code:

This version of the code implements BFS to search for the target file by exploring all files and directories in the current directory before moving on to the subdirectories. The BFS algorithm is implemented using a queue that holds the directories to be processed.